+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/search/MigrateSearchViewModel.kt
@@ -1,0 +1,62 @@
+package eu.kanade.tachiyomi.ui.browse.migration.search
+
+import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.viewmodel.CreationExtras
+import androidx.lifecycle.viewmodel.initializer
+import androidx.lifecycle.viewmodel.viewModelFactory
+import eu.kanade.domain.source.service.SourcePreferences
+import eu.kanade.tachiyomi.source.Source
+import eu.kanade.tachiyomi.ui.browse.source.globalsearch.SearchItemResult
+import eu.kanade.tachiyomi.ui.browse.source.globalsearch.SearchViewModel
+import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
+import tachiyomi.domain.manga.interactor.GetManga
+import tachiyomi.domain.source.service.SourceManager
+import uy.kohesive.injekt.Injekt
+import uy.kohesive.injekt.api.get
+
+class MigrateSearchViewModel(
+    val mangaId: Long,
+    getManga: GetManga = Injekt.get(),
+    private val sourceManager: SourceManager = Injekt.get(),
+    private val sourcePreferences: SourcePreferences = Injekt.get(),
+) : SearchViewModel() {
+
+    companion object {
+        val MANGA_ID_KEY = CreationExtras.Key<Long>()
+
+        val Factory = viewModelFactory {
+            initializer {
+                MigrateSearchViewModel(
+                    mangaId = get(MANGA_ID_KEY)!!,
+                )
+            }
+        }
+    }
+
+    private val migrationSources by lazy { sourcePreferences.migrationSources.get() }
+
+    override val sortComparator = { map: Map<Source, SearchItemResult> ->
+        compareBy<Source>(
+            { (map[it] as? SearchItemResult.Success)?.isEmpty ?: true },
+            { migrationSources.indexOf(it.id) },
+        )
+    }
+
+    init {
+        viewModelScope.launch {
+            val manga = getManga.await(mangaId)!!
+            mutableState.update {
+                it.copy(
+                    from = manga,
+                    searchQuery = manga.title,
+                )
+            }
+            search()
+        }
+    }
+
+    override fun getEnabledSources(): List<Source> {
+        return migrationSources.mapNotNull { sourceManager.get(it) }
+    }
+}