+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/source/SourcesFilterScreen.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.platform.LocalContext
-import cafe.adriel.voyager.core.model.rememberScreenModel
+import androidx.lifecycle.viewmodel.compose.viewModel
 import cafe.adriel.voyager.navigator.LocalNavigator
 import cafe.adriel.voyager.navigator.currentOrThrow
 import eu.kanade.presentation.browse.SourcesFilterScreen
@@ -19,15 +19,15 @@ class SourcesFilterScreen : Screen() {
     @Composable
     override fun Content() {
         val navigator = LocalNavigator.currentOrThrow
-        val screenModel = rememberScreenModel { SourcesFilterScreenModel() }
-        val state by screenModel.state.collectAsState()
+        val viewModel = viewModel<SourcesFilterViewModel>()
+        val state by viewModel.state.collectAsState()
 
-        if (state is SourcesFilterScreenModel.State.Loading) {
+        if (state is SourcesFilterViewModel.State.Loading) {
             LoadingScreen()
             return
         }
 
-        if (state is SourcesFilterScreenModel.State.Error) {
+        if (state is SourcesFilterViewModel.State.Error) {
             val context = LocalContext.current
             LaunchedEffect(Unit) {
                 context.toast(MR.strings.internal_error)
@@ -36,13 +36,13 @@ class SourcesFilterScreen : Screen() {
             return
         }
 
-        val successState = state as SourcesFilterScreenModel.State.Success
+        val successState = state as SourcesFilterViewModel.State.Success
 
         SourcesFilterScreen(
             navigateUp = navigator::pop,
             state = successState,
-            onClickLanguage = screenModel::toggleLanguage,
-            onClickSource = screenModel::toggleSource,
+            onClickLanguage = viewModel::toggleLanguage,
+            onClickSource = viewModel::toggleSource,
         )
     }
 }